+++ src/Resources/public/js/be.js
@@ -19,7 +19,7 @@ document.addEventListener('DOMContentLoaded', function(event) {
     if (ctrl_ccDescription) {
         let ccmQuestion = document.getElementById('ccmQuestion');
         ccmQuestion.innerText = ctrl_ccDescription.value.length === 0 ? '?' : ctrl_ccDescription.value;
-        ctrl_ccDescription.addEventListener('keyup', function () {
+        ctrl_ccDescription.addEventListener('keydown', function () {
             ccmQuestion.innerText = ctrl_ccDescription.value.length === 0 ? '?' : ctrl_ccDescription.value;
         });
     }
@@ -28,7 +28,7 @@ document.addEventListener('DOMContentLoaded', function(event) {
     if (ctrl_ccAcceptOnce) {
         let ccmButtonOnce = document.getElementById('ccmButtonOnce');
         ccmButtonOnce.innerText = ctrl_ccAcceptOnce.value.length === 0 ? '?' : ctrl_ccAcceptOnce.value;
-        ctrl_ccAcceptOnce.addEventListener('keyup', function () {
+        ctrl_ccAcceptOnce.addEventListener('keydown', function () {
             ccmButtonOnce.innerText = ctrl_ccAcceptOnce.value.length === 0 ? '?' : ctrl_ccAcceptOnce.value;
         });
     }
@@ -37,7 +37,7 @@ document.addEventListener('DOMContentLoaded', function(event) {
     if (ctrl_ccAcceptAlways) {
         let ccmButtonAlways = document.getElementById('ccmButtonAlways');
         ccmButtonAlways.innerText = ctrl_ccAcceptAlways.value.length === 0 ? '?' : ctrl_ccAcceptAlways.value;
-        ctrl_ccAcceptAlways.addEventListener('keyup', function () {
+        ctrl_ccAcceptAlways.addEventListener('keydown', function () {
             ccmButtonAlways.innerText = ctrl_ccAcceptAlways.value.length === 0 ? '?' : ctrl_ccAcceptAlways.value;
         });
     }