+++ CHANGELOG.md
@@ -10,6 +10,12 @@ Types of changes
     Security in case of vulnerabilities.
 )
 
+## [1.1.2](https://github.com/pdir/klaro-consent-manager/tree/1.1.2.) – 2023-03-07
+
+- [Fixed] Some minor bugs in the listing of translations fixed.  
+- [Fixed] Contao custom content elements are now handled correctly by Klaro.  
+  [Changed] The contextual consent event and the watcher update event have been improved.   
+
 ## [1.1.1](https://github.com/pdir/klaro-consent-manager/tree/1.1.1) – 2023-02-20
 
 - [Fixed] Rename google maps template